refactor: ReadNews 도메인 분리
