+++ backend/core/src/main/java/com/rollthedice/backend/domain/news/api/NewsController.java
@@ -2,9 +2,7 @@ package com.rollthedice.backend.domain.news.api;
 
 import com.rollthedice.backend.domain.news.dto.response.NewsDetailResponse;
 import com.rollthedice.backend.domain.news.dto.response.NewsResponse;
-import com.rollthedice.backend.domain.news.dto.response.ReadNewsResponse;
 import com.rollthedice.backend.domain.news.service.NewsService;
-import com.rollthedice.backend.domain.news.service.ReadNewsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -17,7 +15,6 @@ import java.util.List;
 @RequestMapping("news")
 public class NewsController implements NewsApi {
     private final NewsService newsService;
-    private final ReadNewsService readNewsService;
 
     @ResponseStatus(HttpStatus.OK)
     @GetMapping("")
@@ -28,13 +25,8 @@ public class NewsController implements NewsApi {
 
     @ResponseStatus(HttpStatus.OK)
     @GetMapping("/{newsId}")
+    @Override
     public NewsDetailResponse getDetailNews(final @PathVariable Long newsId) {
         return newsService.getDetailNews(newsId);
-    }
-
-    @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/viewed-history")
-    public List<ReadNewsResponse> getReadNews() {
-        return readNewsService.getReadNews();
     }
 }
