+++ backend/core/src/main/java/com/rollthedice/backend/domain/news/repository/ReadNewsRepository.java
@@ -1,7 +1,7 @@
 package com.rollthedice.backend.domain.news.repository;
 
 import com.rollthedice.backend.domain.member.entity.Member;
-import com.rollthedice.backend.domain.news.entity.ReadNews;
+import com.rollthedice.backend.domain.readNews.entity.ReadNews;
 import com.rollthedice.backend.domain.statistics.repository.ReadNewsCustomRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
