+++ backend/core/src/test/java/com/rollthedice/backend/domain/news/api/NewsControllerTest.java
@@ -3,7 +3,7 @@ package com.rollthedice.backend.domain.news.api;
 import com.rollthedice.backend.domain.news.exception.NewsNotFoundException;
 import com.rollthedice.backend.domain.news.repository.NewsRepository;
 import com.rollthedice.backend.domain.news.service.NewsService;
-import com.rollthedice.backend.domain.news.service.ReadNewsService;
+import com.rollthedice.backend.domain.readNews.service.ReadNewsService;
 import com.rollthedice.backend.global.BaseControllerTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -82,7 +82,7 @@ class NewsControllerTest extends BaseControllerTest {
     void getReadNews() throws Exception{
         //when
         final ResultActions perform = mockMvc.perform(
-                get("/news/viewed-history")
+                get("/read-news/viewed-history")
                         .contentType(MediaType.APPLICATION_JSON)
                         .header("Authorization", "Bearer " + accessToken)
         ).andDo(print());
